refactor structure of project
