+++ src/renderer/menu.ts
@@ -5,7 +5,7 @@ import {
   BrowserWindow,
   MenuItemConstructorOptions,
 } from 'electron';
-import { isDev, prodDebug } from './constants';
+import { isDev, prodDebug } from '../shared/constants';
 
 interface DarwinMenuItemConstructorOptions extends MenuItemConstructorOptions {
   selector?: string;
